requirements.txt
asgiref==3.4.0
Django==3.2.4
django-restframework==0.0.1
djangorestframework==3.12.4
gunicorn==20.1.0
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
